Split the `BaseEstimateCalculator` up into multiple files
