+++ Jantimizer/QueryOptimiser/Cost/EstimateCalculators/EstimateCalculatorVariance.cs
@@ -11,9 +11,11 @@ namespace QueryOptimiser.Cost.EstimateCalculators
     internal class EstimateCalculatorVariance : BaseEstimateCalculator
     {
         public override INodeCost<JoinNode> NodeCostCalculator { get; set; }
+        public override MatchFinder<JoinNode> Matcher { get; set; }
 
         public EstimateCalculatorVariance(IHistogramManager manager) : base(manager) {
             NodeCostCalculator = new JoinEstimateEquiDepthVariance();
+            Matcher = new MatchFinder<JoinNode>(NodeCostCalculator);
         }
     }
 }
