+++ Jantimizer/QueryOptimiserTest/Cost/EstimateCalculators/BaseEstimateCalculatorTests.cs
@@ -15,64 +15,6 @@ namespace QueryOptimiserTest.Cost.EstimateCalculators
     [TestClass]
     public class BaseEstimateCalculatorTests
     {
-        #region DoesMatch
-
-        [TestMethod]
-        // Right bucket start index is within Left bucket range
-        // Right Bucket:      |======|
-        // Left Bucket:    |======|
-        [DataRow(0, 10, 5, 20)]
-        [DataRow(0, 1000, 999, 2000)]
-        [DataRow(0, 1000, 500, 1001)]
-        // Right bucket end index is within Left bucket range
-        // Right Bucket:   |======|
-        // Left Bucket:       |======|
-        [DataRow(50, 100, 0, 51)]
-        [DataRow(50, 100, 10, 99)]
-        [DataRow(50, 100, 25, 50)]
-        // Right bucket is entirely within Left bucket
-        // Right Bucket:   |======|
-        // Left Bucket:  |===========|
-        [DataRow(0, 100, 20, 80)]
-        [DataRow(0, 100, 1, 99)]
-        // Left bucket is entirely within Right bucket
-        // Right Bucket: |===========|
-        // Left Bucket:     |=====|
-        [DataRow(20, 80, 0, 100)]
-        [DataRow(1, 99, 0, 100)]
-        public void Can_DoesMatch_TrueCases(int leftStart, int leftEnd, int rightStart, int rightEnd)
-        {
-            // ARRANGE
-            IHistogramBucket leftBucket = new HistogramBucket(leftStart, leftEnd, 1);
-            IHistogramBucket rightBucket = new HistogramBucket(rightStart, rightEnd, 1);
-            var estimator = new EstimateCalculatorEquiDepth(new HistogramManagerStub());
-
-            // ACT
-            var result = estimator.DoesMatch(leftBucket, rightBucket);
-
-            // ASSERT
-            Assert.IsTrue(result);
-        }
-
-        [TestMethod]
-        [DataRow(0, 10, 20, 30)]
-        [DataRow(20, 30, 0, 10)]
-        public void Can_DoesMatch_FalseCases(int leftStart, int leftEnd, int rightStart, int rightEnd)
-        {
-            // ARRANGE
-            IHistogramBucket leftBucket = new HistogramBucket(leftStart, leftEnd, 1);
-            IHistogramBucket rightBucket = new HistogramBucket(rightStart, rightEnd, 1);
-            var estimator = new EstimateCalculatorEquiDepth(new HistogramManagerStub());
-
-            // ACT
-            var result = estimator.DoesMatch(leftBucket, rightBucket);
-
-            // ASSERT
-            Assert.IsFalse(result);
-        }
-
-        #endregion
-
         #region GetOverlappingReferences
 
         [TestMethod]
@@ -347,476 +289,6 @@ namespace QueryOptimiserTest.Cost.EstimateCalculators
             Assert.AreEqual(1, result.RightBuckets.Count);
             Assert.IsTrue(result.LeftBuckets.Contains(tableHistogramBucket1));
             Assert.IsTrue(result.RightBuckets.Contains(tableHistogramBucket2));
-        }
-
-        #endregion
-
-        #region GetMatchBucketIndex
-
-        [TestMethod]
-        [DataRow(
-            new int[] { 1, 11, 21 },
-            new int[] { 10, 20, 30 },
-            0,
-            3,
-            15,
-            1)]
-        [DataRow(
-            new int[] { 1, 11, 21 },
-            new int[] { 10, 20, 30 },
-            0,
-            3,
-            5,
-            0)]
-        [DataRow(
-            new int[] { 1, 11, 21 },
-            new int[] { 10, 20, 30 },
-            0,
-            3,
-            25,
-            2)]
-        public void Can_GetMatchBucketIndex_Found(int[] startIndexes, int[] endIndexes, int lowerBound, int upperBound, IComparable compareValue, int expectedIndex)
-        {
-            // ARRANGE
-            var indexList = new List<IHistogramBucket>();
-            for (int i = 0; i < startIndexes.Length; i++)
-                indexList.Add(new HistogramBucket(startIndexes[i], endIndexes[i], 1));
-
-            var estimator = new EstimateCalculatorEquiDepth(new HistogramManagerStub());
-
-            // ACT
-            var result = estimator.GetMatchBucketIndex(indexList, lowerBound, upperBound, compareValue);
-
-            // ASSERT
-            Assert.AreEqual(expectedIndex, result);
-        }
-
-        [TestMethod]
-        [DataRow(
-            new int[] { 1, 11, 21 },
-            new int[] { 10, 20, 30 },
-            0,
-            3,
-            50,
-            -1)]
-        [DataRow(
-            new int[] { 1, 11, 21 },
-            new int[] { 10, 20, 30 },
-            0,
-            3,
-            0,
-            -1)]
-        public void Can_GetMatchBucketIndex_NotFound(int[] startIndexes, int[] endIndexes, int lowerBound, int upperBound, IComparable compareValue, int expectedIndex)
-        {
-            // ARRANGE
-            var indexList = new List<IHistogramBucket>();
-            for (int i = 0; i < startIndexes.Length; i++)
-                indexList.Add(new HistogramBucket(startIndexes[i], endIndexes[i], 1));
-
-            var estimator = new EstimateCalculatorEquiDepth(new HistogramManagerStub());
-
-            // ACT
-            var result = estimator.GetMatchBucketIndex(indexList, lowerBound, upperBound, compareValue);
-
-            // ASSERT
-            Assert.AreEqual(expectedIndex, result);
-        }
-
-        #endregion
-
-        #region GetBucketBounds
-
-        [TestMethod]
-        [DataRow(
-            new int[] { 1, 11 },
-            new int[] { 10, 20 },
-            new int[] { 5, 31 },
-            new int[] { 30, 40 },
-            2,
-            2)]
-        [DataRow(
-            new int[] { 1, 100 },
-            new int[] { 10, 110 },
-            new int[] { 5, 31 },
-            new int[] { 30, 40 },
-            0,
-            0)]
-        public void Can_GetBucketBounds_Less(int[] leftBucketsStart, int[] leftBucketsEnd, int[] rightBucketsStart, int[] rightBucketsEnd, int expLeftCount, int expRightCount)
-        {
-            // ARRANGE
-            var bucketsLeft = new List<IHistogramBucket>();
-            var rightLeft = new List<IHistogramBucket>();
-            for (int i = 0; i < leftBucketsStart.Length; i++)
-                bucketsLeft.Add(new HistogramBucket(leftBucketsStart[i], leftBucketsEnd[i], 1));
-            for (int i = 0; i < rightBucketsStart.Length; i++)
-                rightLeft.Add(new HistogramBucket(rightBucketsStart[i], rightBucketsEnd[i], 1));
-
-            var estimator = new EstimateCalculatorEquiDepth(new HistogramManagerStub());
-
-            // ACT
-            var result = estimator.GetBucketBounds(ComparisonType.Type.Less, bucketsLeft, rightLeft);
-
-            // ASSERT
-            Assert.AreEqual(expLeftCount, result.LeftBuckets.Count);
-            Assert.AreEqual(expRightCount, result.RightBuckets.Count);
-        }
-
-        [TestMethod]
-        [DataRow(
-            new int[] { 1, 11, 32 },
-            new int[] { 10, 31, 200 },
-            new int[] { 5, 31 },
-            new int[] { 30, 40 },
-            3,
-            2)]
-        [DataRow(
-            new int[] { 1, 11 },
-            new int[] { 10, 31 },
-            new int[] { 1, 31, 41 },
-            new int[] { 10, 40, 500 },
-            2,
-            3)]
-        public void Can_GetBucketBounds_LessOrEqual(int[] leftBucketsStart, int[] leftBucketsEnd, int[] rightBucketsStart, int[] rightBucketsEnd, int expLeftCount, int expRightCount)
-        {
-            // ARRANGE
-            var bucketsLeft = new List<IHistogramBucket>();
-            var rightLeft = new List<IHistogramBucket>();
-            for (int i = 0; i < leftBucketsStart.Length; i++)
-                bucketsLeft.Add(new HistogramBucket(leftBucketsStart[i], leftBucketsEnd[i], 1));
-            for (int i = 0; i < rightBucketsStart.Length; i++)
-                rightLeft.Add(new HistogramBucket(rightBucketsStart[i], rightBucketsEnd[i], 1));
-
-            var estimator = new EstimateCalculatorEquiDepth(new HistogramManagerStub());
-
-            // ACT
-            var result = estimator.GetBucketBounds(ComparisonType.Type.EqualOrLess, bucketsLeft, rightLeft);
-
-            // ASSERT
-            Assert.AreEqual(expLeftCount, result.LeftBuckets.Count);
-            Assert.AreEqual(expRightCount, result.RightBuckets.Count);
-        }
-
-        [TestMethod]
-        [DataRow(
-            new int[] { 1, 11 },
-            new int[] { 10, 20 },
-            new int[] { 21, 31 },
-            new int[] { 30, 40 },
-            0,
-            0)]
-        [DataRow(
-            new int[] { 1, 100 },
-            new int[] { 10, 110 },
-            new int[] { 5, 31 },
-            new int[] { 30, 40 },
-            2,
-            2)]
-        public void Can_GetBucketBounds_More(int[] leftBucketsStart, int[] leftBucketsEnd, int[] rightBucketsStart, int[] rightBucketsEnd, int expLeftCount, int expRightCount)
-        {
-            // ARRANGE
-            var bucketsLeft = new List<IHistogramBucket>();
-            var rightLeft = new List<IHistogramBucket>();
-            for (int i = 0; i < leftBucketsStart.Length; i++)
-                bucketsLeft.Add(new HistogramBucket(leftBucketsStart[i], leftBucketsEnd[i], 1));
-            for (int i = 0; i < rightBucketsStart.Length; i++)
-                rightLeft.Add(new HistogramBucket(rightBucketsStart[i], rightBucketsEnd[i], 1));
-
-            var estimator = new EstimateCalculatorEquiDepth(new HistogramManagerStub());
-
-            // ACT
-            var result = estimator.GetBucketBounds(ComparisonType.Type.More, bucketsLeft, rightLeft);
-
-            // ASSERT
-            Assert.AreEqual(expLeftCount, result.LeftBuckets.Count);
-            Assert.AreEqual(expRightCount, result.RightBuckets.Count);
-        }
-
-        [TestMethod]
-        [DataRow(
-            new int[] { 1, 11, 32 },
-            new int[] { 10, 31, 200 },
-            new int[] { 5, 31 },
-            new int[] { 30, 40 },
-            3,
-            2)]
-        [DataRow(
-            new int[] { 1, 11 },
-            new int[] { 10, 31 },
-            new int[] { 1, 31, 41 },
-            new int[] { 10, 40, 500 },
-            2,
-            2)]
-        public void Can_GetBucketBounds_MoreOrEqual(int[] leftBucketsStart, int[] leftBucketsEnd, int[] rightBucketsStart, int[] rightBucketsEnd, int expLeftCount, int expRightCount)
-        {
-            // ARRANGE
-            var bucketsLeft = new List<IHistogramBucket>();
-            var rightLeft = new List<IHistogramBucket>();
-            for (int i = 0; i < leftBucketsStart.Length; i++)
-                bucketsLeft.Add(new HistogramBucket(leftBucketsStart[i], leftBucketsEnd[i], 1));
-            for (int i = 0; i < rightBucketsStart.Length; i++)
-                rightLeft.Add(new HistogramBucket(rightBucketsStart[i], rightBucketsEnd[i], 1));
-
-            var estimator = new EstimateCalculatorEquiDepth(new HistogramManagerStub());
-
-            // ACT
-            var result = estimator.GetBucketBounds(ComparisonType.Type.EqualOrMore, bucketsLeft, rightLeft);
-
-            // ASSERT
-            Assert.AreEqual(expLeftCount, result.LeftBuckets.Count);
-            Assert.AreEqual(expRightCount, result.RightBuckets.Count);
-        }
-
-        #endregion
-
-        #region MakeNewIntermediateBucket
-
-        [TestMethod]
-        public void Can_MakeNewIntermediateBucket()
-        {
-            // ARRANGE
-            JoinPredicate predicate = new JoinPredicate(
-                new TableReferenceNode(0, "a", "a"),
-                "v",
-                new TableReferenceNode(1, "b", "b"),
-                "v",
-                "a.v > b.v",
-                ComparisonType.Type.More);
-            IHistogramBucket leftBucket = new HistogramBucket(0, 10, 5);
-            IHistogramBucket rightBucket = new HistogramBucket(10, 20, 5);
-
-            var estimator = new EstimateCalculatorEquiDepth(new HistogramManagerStub());
-
-            // ACT
-            var result = estimator.MakeNewIntermediateBucket(predicate, leftBucket, rightBucket);
-
-            // ASSERT
-            Assert.AreEqual(leftBucket, result.Buckets[new TableAttribute("a", "v")].Bucket);
-            Assert.AreEqual(rightBucket, result.Buckets[new TableAttribute("b", "v")].Bucket);
-            Assert.IsTrue(result.Buckets.Keys.Contains(new TableAttribute("a", "v")));
-            Assert.IsTrue(result.Buckets.Keys.Contains(new TableAttribute("b", "v")));
-        }
-
-        #endregion
-
-        #region GetInEqualityMatches
-
-        [TestMethod]
-        [DataRow(
-            // Left buckets
-            new int[] { 1,  11, 21 },
-            new int[] { 10, 20, 30 },
-            // Right buckets
-            new int[] { 1, 6,  31 },
-            new int[] { 5, 30, 50 },
-            4)]
-        [DataRow(
-            // Left buckets
-            new int[] { 1, 3, 5 },
-            new int[] { 2, 4, 6 },
-            // Right buckets
-            new int[] { 1,  11, 21 },
-            new int[] { 10, 20, 30 },
-            3)]
-        public void Can_GetInEqualityMatches_More(int[] leftFrom, int[] leftTo, int[] rightFrom, int[] rightTo, int expIntermediateBucketCount)
-        {
-            // ARRANGE
-            JoinPredicate predicate = new JoinPredicate(
-                new TableReferenceNode(0, "a", "a"),
-                "v",
-                new TableReferenceNode(1, "b", "b"),
-                "v",
-                "a.v > b.v",
-                ComparisonType.Type.More);
-            var bucketsLeft = new List<IHistogramBucket>();
-            var bucketsRight = new List<IHistogramBucket>();
-            for (int i = 0; i < leftFrom.Length; i++)
-                bucketsLeft.Add(new HistogramBucket(leftFrom[i], leftTo[i], 1));
-            for (int i = 0; i < rightFrom.Length; i++)
-                bucketsRight.Add(new HistogramBucket(rightFrom[i], rightTo[i], 1));
-
-            var estimator = new EstimateCalculatorEquiDepth(new HistogramManagerStub());
-
-            // ACT
-            var result = estimator.GetInEqualityMatches(predicate, bucketsLeft, bucketsRight);
-
-            // ASSERT
-            Assert.AreEqual(expIntermediateBucketCount, result.Count);
-        }
-
-        [TestMethod]
-        [DataRow(
-            // Left buckets
-            new int[] { 1, 11, 21 },
-            new int[] { 10, 20, 30 },
-            // Right buckets
-            new int[] { 2, 11, 31 },
-            new int[] { 10, 30, 50 },
-            4)]
-        [DataRow(
-            // Left buckets
-            new int[] { 1, 3, 5 },
-            new int[] { 2, 20, 6 },
-            // Right buckets
-            new int[] { 1, 11, 21 },
-            new int[] { 10, 20, 30 },
-            3)]
-        public void Can_GetInEqualityMatches_MoreOrEqual(int[] leftFrom, int[] leftTo, int[] rightFrom, int[] rightTo, int expIntermediateBucketCount)
-        {
-            // ARRANGE
-            JoinPredicate predicate = new JoinPredicate(
-                new TableReferenceNode(0, "a", "a"),
-                "v",
-                new TableReferenceNode(1, "b", "b"),
-                "v",
-                "a.v >= b.v",
-                ComparisonType.Type.EqualOrMore);
-            var bucketsLeft = new List<IHistogramBucket>();
-            var bucketsRight = new List<IHistogramBucket>();
-            for (int i = 0; i < leftFrom.Length; i++)
-                bucketsLeft.Add(new HistogramBucket(leftFrom[i], leftTo[i], 1));
-            for (int i = 0; i < rightFrom.Length; i++)
-                bucketsRight.Add(new HistogramBucket(rightFrom[i], rightTo[i], 1));
-
-            var estimator = new EstimateCalculatorEquiDepth(new HistogramManagerStub());
-
-            // ACT
-            var result = estimator.GetInEqualityMatches(predicate, bucketsLeft, bucketsRight);
-
-            // ASSERT
-            Assert.AreEqual(expIntermediateBucketCount, result.Count);
-        }
-
-        [TestMethod]
-        [DataRow(
-            // Left buckets
-            new int[] { 1, 11, 21 },
-            new int[] { 10, 20, 30 },
-            // Right buckets
-            new int[] { 2, 11, 31 },
-            new int[] { 10, 30, 50 },
-            3)]
-        [DataRow(
-            // Left buckets
-            new int[] { 1, 3, 5 },
-            new int[] { 2, 20, 6 },
-            // Right buckets
-            new int[] { 1, 11, 21 },
-            new int[] { 10, 20, 30 },
-            9)]
-        public void Can_GetInEqualityMatches_Less(int[] leftFrom, int[] leftTo, int[] rightFrom, int[] rightTo, int expIntermediateBucketCount)
-        {
-            // ARRANGE
-            JoinPredicate predicate = new JoinPredicate(
-                new TableReferenceNode(0, "a", "a"),
-                "v",
-                new TableReferenceNode(1, "b", "b"),
-                "v",
-                "a.v < b.v",
-                ComparisonType.Type.Less);
-            var bucketsLeft = new List<IHistogramBucket>();
-            var bucketsRight = new List<IHistogramBucket>();
-            for (int i = 0; i < leftFrom.Length; i++)
-                bucketsLeft.Add(new HistogramBucket(leftFrom[i], leftTo[i], 1));
-            for (int i = 0; i < rightFrom.Length; i++)
-                bucketsRight.Add(new HistogramBucket(rightFrom[i], rightTo[i], 1));
-
-            var estimator = new EstimateCalculatorEquiDepth(new HistogramManagerStub());
-
-            // ACT
-            var result = estimator.GetInEqualityMatches(predicate, bucketsLeft, bucketsRight);
-
-            // ASSERT
-            Assert.AreEqual(expIntermediateBucketCount, result.Count);
-        }
-
-        [TestMethod]
-        [DataRow(
-            // Left buckets
-            new int[] { 9, 11, 21 },
-            new int[] { 10, 20, 30 },
-            // Right buckets
-            new int[] { 2, 10, 31 },
-            new int[] { 9, 30, 32 },
-            3)]
-        [DataRow(
-            // Left buckets
-            new int[] { 1, 11, 5 },
-            new int[] { 10, 20, 6 },
-            // Right buckets
-            new int[] { 1, 11, 21 },
-            new int[] { 5, 20, 30 },
-            6)]
-        public void Can_GetInEqualityMatches_LessOrEqual(int[] leftFrom, int[] leftTo, int[] rightFrom, int[] rightTo, int expIntermediateBucketCount)
-        {
-            // ARRANGE
-            JoinPredicate predicate = new JoinPredicate(
-                new TableReferenceNode(0, "a", "a"),
-                "v",
-                new TableReferenceNode(1, "b", "b"),
-                "v",
-                "a.v <= b.v",
-                ComparisonType.Type.EqualOrLess);
-            var bucketsLeft = new List<IHistogramBucket>();
-            var bucketsRight = new List<IHistogramBucket>();
-            for (int i = 0; i < leftFrom.Length; i++)
-                bucketsLeft.Add(new HistogramBucket(leftFrom[i], leftTo[i], 1));
-            for (int i = 0; i < rightFrom.Length; i++)
-                bucketsRight.Add(new HistogramBucket(rightFrom[i], rightTo[i], 1));
-
-            var estimator = new EstimateCalculatorEquiDepth(new HistogramManagerStub());
-
-            // ACT
-            var result = estimator.GetInEqualityMatches(predicate, bucketsLeft, bucketsRight);
-
-            // ASSERT
-            Assert.AreEqual(expIntermediateBucketCount, result.Count);
-        }
-
-        #endregion
-
-        #region GetEqualityMatches
-
-        [TestMethod]
-        [DataRow(
-            // Left buckets
-            new int[] { 1, 11, 21 },
-            new int[] { 10, 20, 30 },
-            // Right buckets
-            new int[] { 1, 6, 31 },
-            new int[] { 5, 30, 50 },
-            4)]
-        [DataRow(
-            // Left buckets
-            new int[] { 1, 3, 5 },
-            new int[] { 2, 4, 6 },
-            // Right buckets
-            new int[] { 1, 11, 21 },
-            new int[] { 10, 20, 30 },
-            3)]
-        public void Can_GetEqualityMatches(int[] leftFrom, int[] leftTo, int[] rightFrom, int[] rightTo, int expIntermediateBucketCount)
-        {
-            // ARRANGE
-            JoinPredicate predicate = new JoinPredicate(
-                new TableReferenceNode(0, "a", "a"),
-                "v",
-                new TableReferenceNode(1, "b", "b"),
-                "v",
-                "a.v > b.v",
-                ComparisonType.Type.More);
-            var bucketsLeft = new List<IHistogramBucket>();
-            var bucketsRight = new List<IHistogramBucket>();
-            for (int i = 0; i < leftFrom.Length; i++)
-                bucketsLeft.Add(new HistogramBucket(leftFrom[i], leftTo[i], 1));
-            for (int i = 0; i < rightFrom.Length; i++)
-                bucketsRight.Add(new HistogramBucket(rightFrom[i], rightTo[i], 1));
-
-            var estimator = new EstimateCalculatorEquiDepth(new HistogramManagerStub());
-
-            // ACT
-            var result = estimator.GetEqualityMatches(predicate, bucketsLeft, bucketsRight);
-
-            // ASSERT
-            Assert.AreEqual(expIntermediateBucketCount, result.Count);
         }
 
         #endregion
